线程堵塞问题解决、视角转换

+++ Assets/Script/Individual.cs
@@ -6,11 +6,11 @@ public class Individual
     public int y { get; set; }
     public int z { get; set; }
     public int age { get; set; }
-    public double value { get; set; }
+    public float value { get; set; }
     public int type { get; set; }
-    public double selection { get; set; }
-    public double roulette_left { get; set; }
-    public double roulette_right { get; set; }
+    public float selection { get; set; }
+    public float roulette_left { get; set; }
+    public float roulette_right { get; set; }
 
     public Individual(int xx, int yy, int zz, int type)
     {
